pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "k8s-diagnose"
version = "0.1.0"
description = "K8s 运维诊断智能体 — 对话式排查 Pod 启动失败、调度失败等根因"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0.0",
    "typer>=0.9.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov",
    "ruff>=0.3.0",
    "pyright>=1.1",
]

[project.scripts]
k8s-diagnose = "k8s_diagnose.__main__:main"

[tool.setuptools.packages.find]
include = ["k8s_diagnose*"]

[tool.ruff]
line-length = 120

[tool.pyright]
pythonVersion = "3.11"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
